einj_mem_uc: new test case for copyout
Create a file and write a page of data to it. Map that file and
use an address in the mapped range to inject an error.

Trigger the error by issing a read(2) system call which will
make the kernel copy from the page cache copy of the data (which
has been injected with a UC error).

Signed-off-by: Tony Luck <tony.luck@intel.com>