einj_mem_uc: Fix file descriptor leak for copyin test
The copyin test fails at just over 1000 iterations because it can no
longer open new file descriptors.

The problem is that the test opens a file, and closes at the end of
the function. But since successful recover sends a SIGBUS, the close
is never executed.

Make the file descriptor global and close it in the main() function
after the normal or SIGBUS return paths.

Signed-off-by: Tony Luck <tony.luck@intel.com>