improve testcases from cases/mce/uc/srar2.set3
According to discussions on LKML, multiple mappings in a single process
could result int sending SIGBUS with the incorrect error virutal
address.  This is acceptable because "MCE SRAR on already hwpoisoned
address" is a very rare event, and the impact of bad address could be
negligible, so let this testcase consider that as the expected behavior.