add run_guest.sh
Basically we manually set up test config on the server
to be tested, but that's not enough because we sometimes
want to save serial console output, which could be lost
by kernel crash. So this script sets up test config from
control server which is different from the testing server.
This script is supposed to be called on kvm host server
and the testing server (virtual machine) can be connectable
by ssh without password input.

To properly save serial console output, you need to have
setting that serial console of the target virtual machine
is saved to log file (not character device).