cases/page_migration/mlocked.set3: extend EXPECTED_RETURN_CODE
THP split could be done in two separate manner, defered split and
non-defered split.  Which one is chosen can be unpredictable, so
both cases should be considered as correct behaviors.
When a thp is split in deferred manner, the whole thp (not individual
pages) are migrated together, so the final result is a new thp.