introduce cases/mm/hwpoison/swapin_failure.set3
This testcase requires SystemTap so not handy, so never should
not be used for automated testing.

NOTE: this testcase was written months ago, so that's not supposed
to be run perfectly in the current environment, so keep in mind that
this testcase is to only keep the concept in the repository instead
of some real usage.